api/pyproject.toml
[tool.poetry]
name = "kengpt"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
flask = {extras = ["async"], version = "^2.2.2"}
gunicorn = {extras = ["gevent"], version = "^20.1.0"}
requests = "^2.28.2"
openai = "^0.26.5"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
